+++ CHANGELOG.md
@@ -17,7 +17,7 @@ ChangeLog
 + 【新特性】- 增加计算属性的支持
 + 【新特性】- 增加对常用 SVG 标签的创建支持
 + 【新特性】- 支持 components 属性通过字符串 'self' 引用自身作为子组件
-+ 【新特性】- 组件属性 template / filters / components / computed 支持 static property
++ 【新特性】- 组件属性 template / filters / components / computed / messages 支持 static property
 + 【新特性】- 组件外部声明的 class ／ style 合并到组件的根元素
 + 【新特性】- data.get 方法支持无参获取整个数据
 + 【优化】- 对模板解析结果的结构进行了优化：